Add lnnode vendor switch (#431)
* Add LNNode vendor switch
